fix: adjust memory usage type and improve database query consistency

+++ src/Data/LogData.php
@@ -15,7 +15,7 @@ final class LogData
         public Response $response,
         public Carbon $sentAt,
         public int|float $durationMs,
-        public float $memoryUsage,
+        public int $memoryUsage,
         public string $fingerprint,
         public array $customFields
     ) {
